app/requirements.txt
Flask==2.0.2
mysql-connector-python
requests==2.28.1
flask-cors
pandas
